requirements.txt
ag2==0.7.3
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fast-depends==2.4.12
fix-busted-json==0.0.18
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mypy-extensions==1.0.0
numpy==2.2.2
ollama==0.4.7
openai==1.61.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pyautogen==0.7.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pywin32==308
pywinauto==0.6.9
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
